+++ mpf/src/july19Forward.py
@@ -4,10 +4,10 @@ import pyexcel_ods
 from datetime import datetime
 import math
 import numpy
+import utils
 import pathlib
 import sys
 import os
-nan = float('nan')
 
 def read_data(path):
     path = './src/' + path
@@ -19,65 +19,10 @@ def read_data(path):
                          path + '. O seguinte erro foi gerado: ' + excep)
         os._exit(1)
 
-# Strange way to check nan. Only I managed to make work
-# Source: https://stackoverflow.com/a/944712/5822594
-def isNaN(string):
-    return string != string
-
-def get_begin_row(rows, begin_string):
-    begin_row = 0
-    for row in rows:
-        begin_row += 1
-        if isinstance(row[0], str) and begin_string in row[0]:
-            break
-    # We need to continue interate until wee a value that is not
-    # whitespace. That happen due to the spreadsheet formatting.
-    while isNaN(rows[begin_row][0]):
-        begin_row += 1
-    return begin_row
-
-def get_end_row(rows, begin_row, end_string):
-    end_row = 0
-    for row in rows:
-        # First goes to begin_row.
-        if end_row < begin_row:
-            end_row += 1
-            continue
-        # Then keep moving until find a blank row.
-        if isinstance(row[0], str) and end_string in row[0]:
-            break
-        if isNaN(row[0]):
-            break
-        end_row += 1
-    return end_row
-
-def type_employee(fn):
-    if 'membros' in fn:
-        return 'membro'
-    if 'servidores' in fn:
-        return 'servidor'
-    if 'pensionistas' in fn:
-        return 'pensionista'
-    if 'colaboradores' in fn:
-        return 'colaborador'
-    raise ValueError('Tipo de inválido de funcionário público: ' + fn)
-
-def treat_rows(rows): 
-  emps_clean = []
-  begin_string = "Matrícula"
-  end_string = "TOTAL GERAL"
-  begin_row = get_begin_row(rows, begin_string)
-  end_row = get_end_row(rows, begin_row, end_string)
-  for row in rows:
-    emp_clean = [x for x in row if str(x) != 'nan']
-    emps_clean.append(emp_clean)
-  return emps_clean[begin_row:end_row]
-
-
 def parse_employees(file_name):
     rows = read_data(file_name).to_numpy()
-    emps_clean = treat_rows(rows)
-    typeE = type_employee(file_name)
+    emps_clean = utils.treat_rows(rows)
+    typeE = utils.type_employee(file_name)
     activeE = 'inativos' not in file_name and 'pensionistas' not in file_name
     employees = {}
     curr_row = 0
@@ -123,7 +68,7 @@ def parse_employees(file_name):
 
 def update_employee_indemnity(file_name, employees):
     rows = read_data(file_name).to_numpy()
-    emp_idemnity = treat_rows(rows)
+    emp_idemnity = utils.treat_rows(rows)
     for row in emp_idemnity:
         emp = employees[row[0]]
         emp['income']['perks'].update({
@@ -159,8 +104,6 @@ def update_employee_indemnity(file_name, employees):
         employees[row[0]] = emp
 
     return employees
-
-
 
 def parse(file_names):
     employees = {}
